+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.4 Feature hooks, some fixes (26 Oct. 2017)
+
+- Added BeforeFeature/AfterFeature hooks
+- Added currentFeature to currentRun state
+- testRunner can pipe inner-outputs to main process, dependent on env setting
+
 ## 0.0.2, 0.0.3 Minor fixes (25 Oct. 2017)
 
 - Fixes to package.json git links